add allow url list

+++ post/views.py
@@ -7,12 +7,18 @@ from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib import messages
 
-def index(request):
-    posts = Post.objects.all()
-    return render(request, 'post/post_list.html', {'posts': posts})
+
+ALLOW_URL_LIST = settings.ALLOW_URL_LIST
 
 
 def post_list(request):
+    
+    # URL CHECK
+    client_ip = request.META.get('REMOTE_ADDR')
+    
+    if client_ip not in ALLOW_URL_LIST:
+        return redirect('account:urlerror') 
+    
     posts = Post.objects.all()
     search_key = request.GET.get("keyword")
     
@@ -22,11 +28,18 @@ def post_list(request):
     return render(request, 'post/post_list.html', {'posts': posts, 'q': search_key})
  
 def index(request):
+    # URL CHECK
+    client_ip = request.META.get('REMOTE_ADDR')
+
+    if client_ip not in ALLOW_URL_LIST:
+        return redirect('account:urlerror') 
+    
     posts = Post.objects.all()
     search_key = request.GET.get("keyword", "")
-    print(search_key)
+
     if search_key:
         print(search_key)
+
         posts = Post.objects.filter(title__icontains=search_key)
        
     return render(request, 'post/post_list.html', {'posts':posts, 'q':search_key})
@@ -36,6 +49,12 @@ def index(request):
     # return render(request, 'post/post_list.html', {'posts': posts})
  
 def post_detail(request, pk):
+    # URL CHECK
+    client_ip = request.META.get('REMOTE_ADDR')
+
+    if client_ip not in ALLOW_URL_LIST:
+        return redirect('account:urlerror') 
+    
     post = get_object_or_404(Post, pk=pk)
     comments = post.comment_set.all()  # 댓글을 가져오는 부분을 수정
     return render(request, 'post/post_detail.html', {'post': post, 'comments': comments})
@@ -43,6 +62,12 @@ def post_detail(request, pk):
 
 @login_required
 def post_new(request):
+    # URL CHECK
+    client_ip = request.META.get('REMOTE_ADDR')
+
+    if client_ip not in ALLOW_URL_LIST:
+        return redirect('account:urlerror') 
+    
     if request.method == 'POST':
         form = PostForm(request.POST)
         if form.is_valid():
@@ -55,14 +80,28 @@ def post_new(request):
         form = PostForm()
     return render(request, 'post/post_form.html', {'form': form})
  
+ 
 def post_delete(request, pk):
+    # URL CHECK
+    client_ip = request.META.get('REMOTE_ADDR')
+
+    if client_ip not in ALLOW_URL_LIST:
+        return redirect('account:urlerror') 
+
     post = get_object_or_404(Post, pk=pk)
     if request.method == 'POST':
         post.delete()
         return redirect('post:post_list')
     return render(request, 'post/post_delete.html', {'post': post})
  
+ 
 def post_edit(request, pk):
+    # URL CHECK
+    client_ip = request.META.get('REMOTE_ADDR')
+
+    if client_ip not in ALLOW_URL_LIST:
+        return redirect('account:urlerror') 
+    
     post = get_object_or_404(Post, pk=pk)
    
     if request.method == 'GET':
@@ -79,17 +118,38 @@ def post_edit(request, pk):
     })
 
 def Comment2(request, pk):
+    # URL CHECK
+    client_ip = request.META.get('REMOTE_ADDR')
+
+    if client_ip not in ALLOW_URL_LIST:
+        return redirect('account:urlerror') 
+
     post = get_object_or_404(Post, pk=pk)
-    comments = CommentForm(request.POST)
-    if comments.is_valid():
-        comments = comments.save(commit=False)
-        comments.post = get_object_or_404(Post, pk=pk)
-        comments.user = request.user
-        comments.save()
+    
+    if request.method == 'POST':
+        comments_form = CommentForm(request.POST)
+        
+        if comments_form.is_valid():
+            comment = comments_form.save(commit=False)
+            comment.post = post
+            comment.user = request.user
+            comment.save()
+            
+            return redirect('post:post_detail', pk=pk)
+    else:
+        # GET 요청이거나 유효하지 않은 데이터가 POST로 전송된 경우의 처리를 여기에 추가할 수 있습니다.
+        messages.error(request, '유효하지 않은 요청입니다.')
+        
     return redirect('post:post_detail', pk=pk)
     
 
 def comment_delete(request, pk, comment_id):
+    # URL CHECK
+    client_ip = request.META.get('REMOTE_ADDR')
+
+    if client_ip not in ALLOW_URL_LIST:
+        return redirect('account:urlerror') 
+
     if request.method == 'POST':
         post = get_object_or_404(Post, pk=pk)
